remove @everyone ping
The default of pinging everyone annoyed my roommates. This change lets
people set the notification level for the channel to "All messages" if
they choose to opt-in to AQI notifications.

This makes the channel be the unit of notification verbosity; another
option would be to create a role that is pinged instead. This was not
done due to complexity.